requirements.txt
# Django and REST Framework
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Configuration
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-yasg==1.21.7

# Filtering
django-filter==23.5

# Utilities
Pillow==10.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Development
django-debug-toolbar==4.2.0
